+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## Changelog
 
-0.3.1 - 0.3.2 (2/17/26)
+0.3.1 - 0.3.3 (2/17/26)
 
 - Fixed it so Arrays of custom types are now properly created/referenced.
 - Missed a replacement function in the C# builder (0.3.2)
+- I'm _SOOO_ stupid. Zod now does all properties (0.3.3). 
+  - Also fixed the default ordering for Zod types.
 
 0.3.0 (2/16/26)